requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==1.10.8
pydantic-settings==2.1.0
sqlalchemy==2.0.27
pymysql==1.1.0
cryptography==42.0.2
email-validator==2.1.0.post1
bcrypt==4.0.1
passlib==1.7.4
python-jose[cryptography]==3.3.0 
xrpl==2.10.0